Take An Optional List_ID In Config.
Since the list_id is simply set by looking up the id using the name of
the list and api, add a config param that can set this automatically. If
this isn't set, it defaults to nil causing the list_id to be set
via the original method.

GitHub:
issue #41

Conflicts:
	spec/lib/list_interface_spec.rb