Remove All Hash Params In List Interface
Mailchimp's new V2 API removes the need for using one big hash as a
param. However, they IMHO take a serious step backwards and move to a
specific ordered set of params. Remove the use of Hashes in the list
interface and set each call to match the correct order of parameters.

GitHub:
issue #41