Clean Gemfile Up
No to libnotify, no no no no no.
Bump rspec rails to just shy of 3. I could have sworn this was done
previously but maybe it didn't bleed from master to another branch. It
was directly patched on that branch. Feel free to use this commit as a
cherry pick point for any branches that need this bump