Improve user sync (#87)
This PR attempts to improve the user performace from Mailchimp.  The
diff covers the following improvements:

 * Speed: instead of iterating through Spree users one by one and
looking them up in mailchimp, prefer to pull all subscription data in a
big batch from Mailchimp, and then update local db as needed.
 * User-friendliness: while I preserved exact functionality, I
renamed/scoped the task to more clearly reflect the directionality of
this sync.  In particular, based on comments and lack of documentation,
it was unclear what this task was actually doing.  Now we specify
explicitly in the README and code that we're pulling subscription status
from Mailchimp.